Finished pick-up script
A functional script for "picking up" parts of the puzzle, that works by raycasting from the camera, turning off the hit puzzle piece, turning on a fake puzzle piece in the player's "Hand" to make it look like it's being held, and then raycasting again to put it down. The putting it down part is not yet included, but this needs to be published so the person setting up the puzzle logic script can connect the two.